requirements.txt
aniso8601==8.0.0
astroid==2.3.3
autopep8==1.4.3
beautifulsoup4==4.8.1
boto3==1.10.26
botocore==1.13.26
cachetools==3.1.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
cssmin==0.2.0
docutils==0.15.2
Flask==1.0.2
Flask-Assets==0.12
Flask-Cors==3.0.7
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-RESTful==0.3.7
Flask-WTF==0.14.2
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-vision==0.40.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsmin==2.2.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.0
mccabe==0.6.1
peewee==3.8.2
peewee-db-evolve==3.7.0
protobuf==3.10.0
psycopg2-binary==2.7.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2019.3
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
soupsieve==1.9.5
typed-ast==1.4.0
urllib3==1.25.7
webassets==0.12.1
Werkzeug==0.16.0
wikipedia==1.4.0
wrapt==1.11.2
WTForms==2.2.1
